+++ packages/webawesome/src/components/color-picker/color-picker.ts
@@ -276,7 +276,7 @@ export default class WaColorPicker extends WebAwesomeFormAssociatedElement {
     this.format = formats[nextIndex] as 'hex' | 'rgb' | 'hsl' | 'hsv';
     this.setColor(this.value || '');
     this.dispatchEvent(new Event('change', { bubbles: true, composed: true }));
-    this.dispatchEvent(new InputEvent('input'));
+    this.dispatchEvent(new InputEvent('input', { bubbles: true, composed: true }));
   }
 
   private handleAlphaDrag(event: PointerEvent) {
@@ -296,7 +296,7 @@ export default class WaColorPicker extends WebAwesomeFormAssociatedElement {
 
         if (this.value !== currentValue) {
           currentValue = this.value;
-          this.dispatchEvent(new InputEvent('input'));
+          this.dispatchEvent(new InputEvent('input', { bubbles: true, composed: true }));
         }
       },
       onStop: () => {
@@ -326,7 +326,7 @@ export default class WaColorPicker extends WebAwesomeFormAssociatedElement {
 
         if (this.value !== currentValue) {
           currentValue = this.value;
-          this.dispatchEvent(new InputEvent('input'));
+          this.dispatchEvent(new InputEvent('input', { bubbles: true, composed: true }));
         }
       },
       onStop: () => {
@@ -359,7 +359,7 @@ export default class WaColorPicker extends WebAwesomeFormAssociatedElement {
 
         if (this.value !== currentValue) {
           currentValue = this.value;
-          this.dispatchEvent(new InputEvent('input'));
+          this.dispatchEvent(new InputEvent('input', { bubbles: true, composed: true }));
         }
       },
       onStop: () => {
@@ -402,7 +402,7 @@ export default class WaColorPicker extends WebAwesomeFormAssociatedElement {
     }
 
     if (this.value !== oldValue) {
-      this.dispatchEvent(new InputEvent('input'));
+      this.dispatchEvent(new InputEvent('input', { bubbles: true, composed: true }));
       this.dispatchEvent(new Event('change', { bubbles: true, composed: true }));
     }
   }
@@ -436,7 +436,7 @@ export default class WaColorPicker extends WebAwesomeFormAssociatedElement {
     }
 
     if (this.value !== oldValue) {
-      this.dispatchEvent(new InputEvent('input'));
+      this.dispatchEvent(new InputEvent('input', { bubbles: true, composed: true }));
       this.dispatchEvent(new Event('change', { bubbles: true, composed: true }));
     }
   }
@@ -470,7 +470,7 @@ export default class WaColorPicker extends WebAwesomeFormAssociatedElement {
     }
 
     if (this.value !== oldValue) {
-      this.dispatchEvent(new InputEvent('input'));
+      this.dispatchEvent(new InputEvent('input', { bubbles: true, composed: true }));
       this.dispatchEvent(new Event('change', { bubbles: true, composed: true }));
     }
   }
@@ -490,7 +490,7 @@ export default class WaColorPicker extends WebAwesomeFormAssociatedElement {
     }
 
     if (this.value !== oldValue) {
-      this.dispatchEvent(new InputEvent('input'));
+      this.dispatchEvent(new InputEvent('input', { bubbles: true, composed: true }));
       this.dispatchEvent(new Event('change', { bubbles: true, composed: true }));
     }
   }
@@ -511,7 +511,7 @@ export default class WaColorPicker extends WebAwesomeFormAssociatedElement {
         this.input.value = this.value;
 
         if (this.value !== oldValue) {
-          this.dispatchEvent(new InputEvent('input'));
+          this.dispatchEvent(new InputEvent('input', { bubbles: true, composed: true }));
           this.dispatchEvent(new Event('change', { bubbles: true, composed: true }));
         }
 
@@ -696,7 +696,7 @@ export default class WaColorPicker extends WebAwesomeFormAssociatedElement {
         this.setColor(colorSelectionResult.sRGBHex);
 
         if (this.value !== oldValue) {
-          this.dispatchEvent(new InputEvent('input'));
+          this.dispatchEvent(new InputEvent('input', { bubbles: true, composed: true }));
           this.dispatchEvent(new Event('change', { bubbles: true, composed: true }));
         }
       })
@@ -712,7 +712,7 @@ export default class WaColorPicker extends WebAwesomeFormAssociatedElement {
       this.setColor(color);
 
       if (this.value !== oldValue) {
-        this.dispatchEvent(new InputEvent('input'));
+        this.dispatchEvent(new InputEvent('input', { bubbles: true, composed: true }));
         this.dispatchEvent(new Event('change', { bubbles: true, composed: true }));
       }
     }