+++ packages/webawesome/src/components/input/input.ts
@@ -233,7 +233,7 @@ export default class WaInput extends WebAwesomeFormAssociatedElement {
     if (this.value !== '') {
       this.value = '';
       this.dispatchEvent(new WaClearEvent());
-      this.dispatchEvent(new InputEvent('input'));
+      this.dispatchEvent(new InputEvent('input', { bubbles: true, composed: true }));
       this.dispatchEvent(new Event('change', { bubbles: true, composed: true }));
     }