+++ packages/webawesome/src/components/radio-group/radio-group.ts
@@ -171,7 +171,7 @@ export default class WaRadioGroup extends WebAwesomeFormAssociatedElement {
     }
 
     if (this.value !== oldValue) {
-      this.dispatchEvent(new InputEvent('input'));
+      this.dispatchEvent(new InputEvent('input', { bubbles: true, composed: true }));
       this.dispatchEvent(new Event('change', { bubbles: true, composed: true }));
     }
   };
@@ -274,7 +274,7 @@ export default class WaRadioGroup extends WebAwesomeFormAssociatedElement {
     }
 
     if (this.value !== oldValue) {
-      this.dispatchEvent(new InputEvent('input'));
+      this.dispatchEvent(new InputEvent('input', { bubbles: true, composed: true }));
       this.dispatchEvent(new Event('change', { bubbles: true, composed: true }));
     }