+++ packages/webawesome/docs/docs/resources/changelog.md
@@ -66,6 +66,7 @@ During the alpha period, things might break! We take breaking changes very serio
 - Fixed the tooltip position in `<wa-slider>` when using RTL
 - Fixed a bug in `<wa-details>` and native `<details>` styles that made the summary hard to click [issue:684]
 - Fixed a handful of bugs unify form control height across components and native elements
+- Fixed a bug where `input` events from components weren't bubbling
 - Improved CSS utilities and Native Styles to use [CSS layers](https://developer.mozilla.org/en-US/docs/Web/CSS/@layer) for easier end user customization (no more specificity conflicts — your CSS wins!)
 - Improved native `<button>` styles to properly space icons
 - Improved button appearances in `<wa-color-picker>`